Move 'mathiness' classes into lane.h (the only place where they are used).
