+++ drake/automotive/maliput/monolane/CMakeLists.txt
@@ -6,7 +6,6 @@ drake_install_headers(
   junction.h
   lane.h
   line_lane.h
-  mathiness.h
   road_geometry.h
   segment.h
   )
