+++ drake/automotive/maliput/monolane/segment.h
@@ -8,8 +8,6 @@
 #include "drake/automotive/maliput/api/lane.h"
 #include "drake/automotive/maliput/api/segment.h"
 
-#include "drake/automotive/maliput/monolane/mathiness.h"
-
 namespace drake {
 namespace maliput {
 namespace monolane {
