+++ drake/automotive/maliput/monolane/test/CMakeLists.txt
@@ -4,9 +4,6 @@ target_link_libraries(hodgepodge drakeAutomotive)
 drake_add_cc_test(monolane_lanes_test)
 target_link_libraries(monolane_lanes_test drakeAutomotive)
 
-drake_add_cc_test(monolane_mathiness_test)
-target_link_libraries(monolane_mathiness_test drakeAutomotive)
-
 drake_add_cc_test(monolane_builder_test)
 target_link_libraries(monolane_builder_test drakeAutomotive)
